Move validation stuff to new package `validate`
