+++ ead/validate/validate.go
@@ -1,4 +1,4 @@
-package ead
+package validate
 
 func ValidateEAD(ead []byte) []string {
 	return []string{"RED LIGHT"}
